# Optimized row/col getindex for a `BandedMatrix`