```julia
julia> B = brand(Int, 200, 200, 20, 20);

julia> @btime $B[BandRange,1];
  53.515 ns (1 allocation: 224 bytes) # PR
  101.435 ns (1 allocation: 224 bytes) # master

julia> @btime $B[1,BandRange];
  75.047 ns (1 allocation: 224 bytes) # PR
  99.042 ns (1 allocation: 224 bytes) # master
```
Similarly, with a `Colon` as one of the indices. The improvement comes from forwarding the indexing to the data.